Upgrade JSON schema validator
This seems to have introduced better evaluation of `allOf` and `oneOf` references.

However, it does seem to introduce a bug which causes it not to consider properties in nested schemas more than 1 level deep when evaluating `unevaluatedProperties`: https://github.com/networknt/json-schema-validator/issues/1123

Therefore, a hack was introduced to make sure validation succeeds.. Note that not all concrete types are currently evaluated (#404), so there are many other schemas which would also fail if they were to be evaluated, and they would need to apply a similar fix.